fix: production-readiness audit — 10 bugs across 7 files
🔴 Critical:
- QuoteCreation: stale item.totaal on save/PDF — recalculate at save time
- QuoteCreation: project_id "geen" saved as FK — guard with !== 'geen'
- WerkbonDetail: hardcoded BTW 21% in totalen + factuur conversion — use standaardBtw from AppSettings
- WerkbonDetail: factuur nummer collision (Date.now slice) — sequential FAC-YYYY-NNN
- TijdregistratieLayout: hardcoded BTW 0.21 + non-unique factuur nummer — use standaardBtw + sequential nummer
- ProjectDetail: factuur nummer collision (Date.now slice) — sequential FAC-YYYY-NNN

🟠 Medium:
- WerkbonDetail: handleIndienen race condition on new werkbonnen — block submit until saved
- BestelbonDetail: displayed subtotaal excl BTW but saved totaal incl BTW — show both
- vercel.json: CORS Access-Control-Allow-Origin: * with auth headers — removed wildcard

🟡 Low:
- QuoteItemsTable: parseInt truncates decimal BTW rates — changed to parseFloat

+++ workmate/src/components/quotes/QuoteItemsTable.tsx
@@ -333,7 +333,7 @@ export function QuoteItemsTable({
                       <label className="text-[11px] font-medium text-muted-foreground">BTW</label>
                       <Select
                         value={String(item.btw_percentage)}
-                        onValueChange={(value) => onUpdateItem(item.id, 'btw_percentage', parseInt(value))}
+                        onValueChange={(value) => onUpdateItem(item.id, 'btw_percentage', parseFloat(value))}
                       >
                         <SelectTrigger className="h-9 text-sm">
                           <SelectValue />
